requirements.txt
instabot==0.32.0
requests==2.21.0
python-dotenv==0.9.1